LiveViewerPlot.py: moved module into LiveViewer package
